pyproject.toml
[project]
name = "dviforbml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "hydra-submitit-launcher>=1.2.0",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "numpy>=2.2.3",
    "pandas>=2.2.3",
    "scipy>=1.15.2",
    "seaborn>=0.13.2",
    "torch>=2.4.1",
    "torch-directml>=0.2.5.dev240914",
    "tqdm>=4.67.1",
    "types-seaborn>=0.13.2.20250111",
    "types-tqdm>=4.67.0.20250301",
    "wandb>=0.19.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dviforbml"]